fix css

+++ diaries/static/js/del.js
@@ -1,36 +1,38 @@
 function deleteDiary(diaryId) {
-    const csrftoken = getCookie('csrftoken');
+  const csrftoken = getCookie("csrftoken");
+
+  if (confirm("Are you sure you want to delete this diary entry?")) {
     fetch(`/delete-diary/${diaryId}/`, {
-        method: 'DELETE',
-        headers: {
-            'X-CSRFToken': csrftoken,
-            'X-Requested-With': 'XMLHttpRequest'
-        }
+      method: "DELETE",
+      headers: {
+        "X-CSRFToken": csrftoken,
+        "X-Requested-With": "XMLHttpRequest",
+      },
     })
-    .then(response => response.json())
-    .then(data => {
+      .then((response) => response.json())
+      .then((data) => {
         if (data.message) {
-            alert(data.message);
-            location.reload()
-            // document.getElementById(`diary-${diaryId}`).remove();
+          alert(data.message);
+          location.reload();
         } else {
-            alert('An error occurred.');
+          alert("An error occurred.");
         }
-    })
-    .catch(error => console.error('Error:', error));
+      })
+      .catch((error) => console.error("Error:", error));
+  }
 }
 
 function getCookie(name) {
-    let cookieValue = null;
-    if (document.cookie && document.cookie !== '') {
-        const cookies = document.cookie.split(';');
-        for (let i = 0; i < cookies.length; i++) {
-            const cookie = cookies[i].trim();
-            if (cookie.substring(0, name.length + 1) === (name + '=')) {
-                cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-                break;
-            }
-        }
+  let cookieValue = null;
+  if (document.cookie && document.cookie !== "") {
+    const cookies = document.cookie.split(";");
+    for (let i = 0; i < cookies.length; i++) {
+      const cookie = cookies[i].trim();
+      if (cookie.substring(0, name.length + 1) === name + "=") {
+        cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
+        break;
+      }
     }
-    return cookieValue;
+  }
+  return cookieValue;
 }
